# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 買いどきナビ API
  description: |
    価格変動を監視し、値上げ・値下げをユーザーに通知するアプリケーションのAPI
    
    ## 主な機能
    - 商品の価格情報取得
    - 価格履歴の照会
    - 通知設定の管理
    - お気に入り商品の管理
  version: 1.0.0
  contact:
    name: 買いどきナビ開発チーム
    email: support@kaidoki-navi.example.com

servers:
  - url: https://api.kaidoki-navi.example.com/v1
    description: 本番環境
  - url: https://api-dev.kaidoki-navi.example.com/v1
    description: 開発環境
  - url: http://localhost:3000/v1
    description: ローカル環境

tags:
  - name: products
    description: 商品関連のAPI
  - name: categories
    description: カテゴリ関連のAPI
  - name: favorites
    description: お気に入り関連のAPI
  - name: notifications
    description: 通知設定関連のAPI
  - name: contact
    description: お問い合わせ関連のAPI

paths:
  /products:
    get:
      tags:
        - products
      summary: 商品一覧を取得
      description: |
        商品の一覧を取得します。
        クエリパラメータで検索・フィルタリングが可能です。
      parameters:
        - name: keyword
          in: query
          description: 検索キーワード（商品名で部分一致検索）
          required: false
          schema:
            type: string
          example: "牛乳"
        - name: category
          in: query
          description: カテゴリでフィルタリング
          required: false
          schema:
            type: string
          example: "飲料"
        - name: sort
          in: query
          description: ソート順
          required: false
          schema:
            type: string
            enum: [price_asc, price_desc, name_asc, name_desc, updated_desc]
            default: updated_desc
        - name: limit
          in: query
          description: 取得件数の上限
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: 取得開始位置
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                    description: 検索条件に一致する総商品数
                  limit:
                    type: integer
                  offset:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /products/{productId}:
    get:
      tags:
        - products
      summary: 商品詳細を取得
      description: 指定したIDの商品の詳細情報を取得します
      parameters:
        - name: productId
          in: path
          required: true
          description: 商品ID
          schema:
            type: string
          example: "item-1"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /products/{productId}/price-history:
    get:
      tags:
        - products
      summary: 価格履歴を取得
      description: 指定した商品の価格履歴を取得します
      parameters:
        - name: productId
          in: path
          required: true
          description: 商品ID
          schema:
            type: string
          example: "item-1"
        - name: days
          in: query
          description: 取得する日数
          required: false
          schema:
            type: integer
            enum: [7, 30, 60, 90, 180]
            default: 30
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  productId:
                    type: string
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/PriceHistory'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /categories:
    get:
      tags:
        - categories
      summary: カテゴリ一覧を取得
      description: 商品カテゴリの一覧を取得します
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /favorites:
    get:
      tags:
        - favorites
      summary: お気に入り商品一覧を取得
      description: ユーザーのお気に入り商品一覧を取得します
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorites:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - favorites
      summary: お気に入りに追加
      description: 商品をお気に入りに追加します
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
              properties:
                productId:
                  type: string
                  description: 商品ID
                  example: "item-1"
      responses:
        '201':
          description: 追加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "お気に入りに追加しました"
                  productId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 既にお気に入りに追加済み
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /favorites/{productId}:
    delete:
      tags:
        - favorites
      summary: お気に入りから削除
      description: 商品をお気に入りから削除します
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          description: 商品ID
          schema:
            type: string
          example: "item-1"
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "お気に入りから削除しました"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /notifications/settings:
    get:
      tags:
        - notifications
      summary: 通知設定を取得
      description: ユーザーの通知設定を取得します
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - notifications
      summary: 通知設定を更新
      description: ユーザーの通知設定を更新します
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettings'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "通知設定を更新しました"
                  settings:
                    $ref: '#/components/schemas/NotificationSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /notifications/line/connect:
    post:
      tags:
        - notifications
      summary: LINE連携
      description: LINEアカウントと連携します
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - lineUserId
              properties:
                lineUserId:
                  type: string
                  description: LINE User ID
                  example: "U1234567890abcdef"
      responses:
        '200':
          description: 連携成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "LINE連携が完了しました"
                  lineUserId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /notifications/line/disconnect:
    post:
      tags:
        - notifications
      summary: LINE連携解除
      description: LINEアカウントとの連携を解除します
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 連携解除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "LINE連携を解除しました"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /contact:
    post:
      tags:
        - contact
      summary: お問い合わせを送信
      description: お問い合わせ内容を送信します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - category
                - message
              properties:
                name:
                  type: string
                  description: お名前
                  example: "山田太郎"
                email:
                  type: string
                  format: email
                  description: メールアドレス
                  example: "yamada@example.com"
                category:
                  type: string
                  enum: [service, technical, privacy, other]
                  description: お問い合わせ種別
                  example: "service"
                message:
                  type: string
                  description: お問い合わせ内容
                  example: "サービスの使い方について質問があります。"
      responses:
        '200':
          description: 送信成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "お問い合わせを受け付けました"
                  contactId:
                    type: string
                    description: お問い合わせID
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン

  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          description: 商品ID
          example: "item-1"
        name:
          type: string
          description: 商品名
          example: "牛乳"
        category:
          type: string
          description: カテゴリ
          example: "飲料"
        currentPrice:
          type: integer
          description: 現在価格（円）
          example: 240
        previousPrice:
          type: integer
          description: 前回価格（円）
          example: 250
        priceChange:
          type: integer
          description: 価格変動（円）
          example: -10
        priceChangePercent:
          type: number
          format: float
          description: 価格変動率（%）
          example: -4.0
        shop:
          type: string
          description: 販売店
          example: "スーパーA"
        lastUpdated:
          type: string
          format: date-time
          description: 最終更新日時
          example: "2025-01-15T10:30:00Z"

    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            description:
              type: string
              description: 商品説明
              example: "新鮮な牛乳です"
            imageUrl:
              type: string
              format: uri
              description: 商品画像URL
              example: "https://example.com/images/milk.jpg"
            priceHistory:
              type: array
              description: 過去30日間の価格履歴
              items:
                $ref: '#/components/schemas/PriceHistory'
            aiSummary:
              type: object
              description: AI分析結果
              properties:
                lowestPrice:
                  type: string
                  example: "直近30日で最安値は230円です"
                trend:
                  type: string
                  example: "来週火曜日に値下げの傾向があります"
                recommendation:
                  type: string
                  example: "今週末の購入がお得です"

    PriceHistory:
      type: object
      properties:
        date:
          type: string
          format: date
          description: 日付
          example: "2025-01-15"
        price:
          type: integer
          description: 価格（円）
          example: 240
        shop:
          type: string
          description: 販売店
          example: "スーパーA"

    Category:
      type: object
      properties:
        id:
          type: string
          description: カテゴリID
          example: "cat-1"
        name:
          type: string
          description: カテゴリ名
          example: "飲料"
        productCount:
          type: integer
          description: このカテゴリの商品数
          example: 25

    NotificationSettings:
      type: object
      properties:
        categories:
          type: array
          description: 通知対象カテゴリ
          items:
            type: string
          example: ["飲料", "生鮮食品"]
        frequency:
          type: string
          enum: [realtime, morning, evening]
          description: 通知頻度
          example: "realtime"
        priceChangeThreshold:
          type: integer
          description: 価格変動の閾値（%）
          example: 5
        lineConnected:
          type: boolean
          description: LINE連携状態
          example: true
        webPushEnabled:
          type: boolean
          description: Web Push通知の有効/無効
          example: false

    Error:
      type: object
      properties:
        error:
          type: string
          description: エラーコード
          example: "BAD_REQUEST"
        message:
          type: string
          description: エラーメッセージ
          example: "リクエストパラメータが不正です"
        details:
          type: array
          description: エラー詳細
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

  responses:
    BadRequest:
      description: リクエストが不正です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "BAD_REQUEST"
            message: "リクエストパラメータが不正です"

    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "UNAUTHORIZED"
            message: "認証トークンが無効です"

    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "NOT_FOUND"
            message: "指定されたリソースが見つかりません"

    InternalServerError:
      description: サーバー内部エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "INTERNAL_SERVER_ERROR"
            message: "サーバー内部でエラーが発生しました"